Re #4: Refactored readings into their own data structure within pcgts (LineReading, part of pageLine.ts), because the previous solution (extra lines) was not really systematic: lines are regions in the source, readings are interpretations of the source.
The PageLine now has additional properties:

* `readings` (a dictionary of the LineReading objects, indexed by readingName)
* `hasReadings`: a switch that broadcasts whether the line has this additional complexity
* `activeReading`: if the line has readings, the name of the reading that is currently active (that is: provided transparently in the `pageLine.sentence` and `.coords`)
* `availableReadings`: an array of the readings for the given pageLIne.
* There is still `transcriptionName`, but this is obsolete now.
* There are methods related to readings, most importantly to set the active reading.

Note that when there are readings, the sentence and coords provided *outside* of the `readings` in the incoming PcGts JSON are stored as a default reading, with the name that is set in the `LineReading.defaultReadingName` static property.

The situation of syllables has also changed: now syllables might exist that are only inside a `LineReading`, not necessarily directly in `textLine.sentence`.

This necessitates a more thorough implementation of (1) searching for syllables by ID (implemented in `PageLine`), and (2) propagating the location of a given syllable within a reading to the Annotation handling section of the application. This is done by the `syllableInfoById(id)` method, which additionally returns the `LineReading` in which the syllable is contained. This is an analogy to `Block.syllableInfoById(id)` -- where this new method of PageLine is called instead of simply `syllableById(id)`.

Finally, the `SyllableConnector` class from `annotations.ts` received an additional field `reading:  LineReading = null`. This is then used in the `annotations-view.component.html` to display only the annotations connected to the active reading of their lines, through a new getter `SyllableConnector.isActive()`.

What needs testing is proper *export* of readings.